add key + change React.memo from List to Card
