+++ src/SimilarBookList.js
@@ -1,4 +1,6 @@
 import React from 'react';
+import SimilarBookCard from "./SimilarBookCard";
+
 
 class SimilarBookList extends React.Component {
     constructor(props){
@@ -43,7 +45,7 @@ class SimilarBookList extends React.Component {
             <div className='w3-row'>
                 {
                     books.map( book  => (
-                <SimilarBookCard removeFromList={this.removeFromList} book={ book }/>
+                <SimilarBookCard key={book.bookId} removeFromList={this.removeFromList} book={ book }/>
                     ))
                 }
             </div>
@@ -51,19 +53,4 @@ class SimilarBookList extends React.Component {
     }
 }
 
-export default React.memo(SimilarBookList);
-
-const SimilarBookCard = ({ book, removeFromList }) => (
-        <div>
-                    <div key={book.bookId} className='w3-col m4 l4 s4'>
-                        <div className='w3-right'>
-                            <button className='w3-button' onClick={() => removeFromList(book.bookId)}>X</button>
-                        </div>
-                        <div className='w3-center w3-large w3-padding'> {book.name}</div>
-                        <div className='w3-left w3-padding'><img alt={book.name} width='70px' height='100px'
-                                                                 src={book.coverImg}/></div>
-                        <div className='w3-small w3-padding'> {book.description} </div>
-                        <div className='w3-right w3-padding'> {book.author} </div>
-                    </div>
-        </div>
-);
+export default SimilarBookList;
